template for toil EC2 instances
This template will create EC2 instances similar to a provisioned scicomp
instance.  Instead of connecting to jumpcloud the intent for this instance
is to share the assigned SSH key pair to allow users to access it.

An alternative to this appraach is to create a jumpbox and leader AMI
and provisioned those from a template.  This will make the template
significantly simpler.