template for toil EC2 instances
This template will create EC2 instances similar to a provisioned scicomp
instance.  The things that this template does not do:
1. does not integrate with jumpcloud
2. does not add anything extra to the AMI you provide it.  It assumes
you've provided an AMI that contains toil and everything it needs which
means you are managing toil jumpbox and leader AMI outside of this
template.

An alternative approach to this is in PR https://github.com/Sage-Bionetworks/toil-cluster-infra/pull/4